pyproject.toml
[project]
name = "sarvam-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.2.2",
    "sarvamai>=0.1.27",
    "streamlit>=1.56.0",
]
